refactor: reorganize ai_edge package structure with src directory
