+++ packages/ai_edge/lib/src/ai_edge_method_channel.dart
