+++ packages/ai_edge/lib/src/types.dart
